Fix platform reboots over IPMI
During platform provisioning, the conductor tries to change the platform
state from `power off` to `rebooting`.
The `reboot` state change in the ipmitool driver has the following sequence:
```
_power_off
_set_boot_device
_power_on
```
.
The conductor sends the ipmitool power off command to the platform that is already powered off.
The platform responds with the message:
```
Set Chassis Power Control to Down/Off failed: Command not supported in present state
```
The provisioning fails because of that.

This patch fixes this behaviour by ignoring the ipmitool return code in this particular flow.